requirements.txt
uvicorn=0.34.0
psycopg2=2.9.10
sqlalchemy=2.0.40
pydantic=2.11.3
fastapi=0.115.12
python_jose=3.4.0
cryptography=44.0.2
bcrypt=4.3.0
cffi=1.17.1
ecdsa=0.19.1
passlib=1.7.4
pyasn1=0.4.8
pycparser=2.22
python-jose=3.4.0
rsa=4.9.1
six=1.17.0
python-multipart=0.0.20